api/requirements.txt
azure-storage-blob
fastapi
pytest
python-jose[cryptography]
python-multipart
requests
uvicorn
